+++ cmd/smyrna/gui/toolboxcallbacks.c
@@ -73,7 +73,7 @@ void btnToolZoom_clicked(GtkWidget * widget, gpointer user_data)
 
 void btnToolZoomIn_clicked(GtkWidget * widget, gpointer user_data)
 {
-	view->zoom = view->zoom + ZOOM_STEP*GetOGLDistance(250);
+	view->zoom = view->zoom + (float)ZOOM_STEP*GetOGLDistance(250);
     if (view->zoom > MAX_ZOOM)
 	view->zoom = (float) MAX_ZOOM;
     glexpose();
@@ -83,7 +83,7 @@ void btnToolZoomIn_clicked(GtkWidget * widget, gpointer user_data)
 void btnToolZoomOut_clicked(GtkWidget * widget, gpointer user_data)
 {
 	view->FontSizeConst=GetOGLDistance(14);
-	view->zoom = view->zoom - ZOOM_STEP*GetOGLDistance(250);
+	view->zoom = view->zoom - (float)ZOOM_STEP*GetOGLDistance(250);
     if (view->zoom < MIN_ZOOM)
 	view->zoom = MIN_ZOOM;
     expose_event(view->drawing_area, NULL, NULL);
